feat: implement EVPN Type 5 path construction in GoBGP runtime
Replace the ErrEVPNNotImplemented stub with a real implementation that
builds and advertises EVPN Type 5 IP Prefix routes for each SRv6
endpoint prefix in a BGPAdvertisement.

- Route distinguisher derived from BGPRouter routerID (Type 1 IP:0)
- MpReachNLRI next-hop is node's primary IPv6 address
- Route target communities parsed from advertisement's communities field
- Support withdrawals via b.DeletePath
- Add configurable BGP_LISTEN_PORT env var for tenant GoBGP instance

+++ internal/cni/cni.go
@@ -125,7 +125,7 @@ func lookupBGPRouter(ctx context.Context, k8s client.Client, nodeName, namespace
 	}
 
 	return bgpConfig{
-		asNumber:   matches[0].Spec.LocalASN,
+		asNumber:   uint32(matches[0].Spec.LocalASN),
 		routerName: matches[0].Name,
 	}, nil
 }
